execution time logger properties moved to package logging
