+++ hw-03-spring-boot/src/main/java/ru/haidarov/hw/config/AppConfig.java
@@ -2,6 +2,7 @@ package ru.haidarov.hw.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
+import ru.haidarov.hw.logging.ExecutionTimeLoggerProperties;
 
 @EnableConfigurationProperties({AppProperties.class, ExecutionTimeLoggerProperties.class})
 @Configuration
